# core/vm: faster* bitvec implementation