--- a/core/vm/analysis.go
+++ b/core/vm/analysis.go
@@ -16,120 +16,26 @@
 
 package vm
 
-const (
-	set2BitsMask = uint16(0b11)
-	set3BitsMask = uint16(0b111)
-	set4BitsMask = uint16(0b1111)
-	set5BitsMask = uint16(0b1_1111)
-	set6BitsMask = uint16(0b11_1111)
-	set7BitsMask = uint16(0b111_1111)
-)
-
 // bitvec is a bit vector which maps bytes in a program.
 // An unset bit means the byte is an opcode, a set bit means
 // it's data (i.e. argument of PUSHxx).
-type bitvec []byte
+type bitvec []uint32
 
-func (bits bitvec) set1(pos uint64) {
-	bits[pos/8] |= 1 << (pos % 8)
+func (bv bitvec) isCode(pc uint64) bool {
+	return (bv[pc/32] & (1 << (pc % 32))) == 0
 }
 
-func (bits bitvec) setN(flag uint16, pos uint64) {
-	a := flag << (pos % 8)
-	bits[pos/8] |= byte(a)
-	if b := byte(a >> 8); b != 0 {
-		bits[pos/8+1] = b
-	}
-}
-
-func (bits bitvec) set8(pos uint64) {
-	a := byte(0xFF << (pos % 8))
-	bits[pos/8] |= a
-	bits[pos/8+1] = ^a
-}
-
-func (bits bitvec) set16(pos uint64) {
-	a := byte(0xFF << (pos % 8))
-	bits[pos/8] |= a
-	bits[pos/8+1] = 0xFF
-	bits[pos/8+2] = ^a
-}
-
-// codeSegment checks if the position is in a code segment.
-func (bits *bitvec) codeSegment(pos uint64) bool {
-	return (((*bits)[pos/8] >> (pos % 8)) & 1) == 0
-}
-
-// codeBitmap collects data locations in code.
-func codeBitmap(code []byte) bitvec {
-	// The bitmap is 4 bytes longer than necessary, in case the code
-	// ends with a PUSH32, the algorithm will set bits on the
-	// bitvector outside the bounds of the actual code.
-	bits := make(bitvec, len(code)/8+1+4)
-	return codeBitmapInternal(code, bits)
-}
-
-// codeBitmapInternal is the internal implementation of codeBitmap.
-// It exists for the purpose of being able to run benchmark tests
-// without dynamic allocations affecting the results.
-func codeBitmapInternal(code, bits bitvec) bitvec {
-	for pc := uint64(0); pc < uint64(len(code)); {
-		op := OpCode(code[pc])
-		pc++
-		if int8(op) < int8(PUSH1) { // If not PUSH (the int8(op) > int(PUSH32) is always false).
-			continue
-		}
-		numbits := op - PUSH1 + 1
-		if numbits >= 8 {
-			for ; numbits >= 16; numbits -= 16 {
-				bits.set16(pc)
-				pc += 16
-			}
-			for ; numbits >= 8; numbits -= 8 {
-				bits.set8(pc)
-				pc += 8
-			}
-		}
-		switch numbits {
-		case 1:
-			bits.set1(pc)
-			pc += 1
-		case 2:
-			bits.setN(set2BitsMask, pc)
-			pc += 2
-		case 3:
-			bits.setN(set3BitsMask, pc)
-			pc += 3
-		case 4:
-			bits.setN(set4BitsMask, pc)
-			pc += 4
-		case 5:
-			bits.setN(set5BitsMask, pc)
-			pc += 5
-		case 6:
-			bits.setN(set6BitsMask, pc)
-			pc += 6
-		case 7:
-			bits.setN(set7BitsMask, pc)
-			pc += 7
-		}
-	}
-	return bits
-}
-
-type bitVec []uint32
-
-func (b bitVec) isCode(pc uint64) bool {
-	return (b[pc/32] & (1 << (pc % 32))) == 0
-}
-
-func newCodeBitVec(code []byte) (bv bitVec) {
-	bv = make(bitVec, len(code)/32+2)
-	bv.codeBitVec(code)
+// newCodeBitVec collects data locations in code.
+func newCodeBitVec(code []byte) (bv bitvec) {
+	bv = make(bitvec, len(code)/32+2)
+	bv.codeBitvecInternal(code)
 	return bv
 }
 
-func (bv bitVec) codeBitVec(code []byte) bitVec {
+// codeBitvecInternal is the internal implementation of codeBitmap.
+// It exists for the purpose of being able to run benchmark tests
+// without dynamic allocations affecting the results.
+func (bv bitvec) codeBitvecInternal(code []byte) bitvec {
 	var pc uint64
 	for pc < uint64(len(code)) {
 		op := code[pc]