… core/vm/testdata/weth9.bytecode