This PR adds a new `bitvec` implementation that is faster in practice and most existing benchmarks.

The core implementation idea is still the same (mark PUSH data with a 1-bit), but by using `[]uint32` instead of `[]byte` the maximum number of slice elements that need to be touched is reduced to two.

For now both implementations exit for easy comparability and testing. Should you decide to merge I would do some cleanup first.

## Benchmark Results
```
$ benchstat -col /v bitvec.bench
                              │     old      │                 new                 │
                              │    sec/op    │   sec/op     vs base                │
JumpdestAnalysis_rand-8         1133.2µ ± 1%   542.6µ ± 1%  -52.11% (p=0.000 n=10)
JumpdestAnalysis_weth9-8        1000.2n ± 0%   890.0n ± 0%  -11.02% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH1-8       821.8µ ± 1%   780.0µ ± 1%   -5.08% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH2-8       588.2µ ± 2%   876.0µ ± 1%  +48.94% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH3-8       463.9µ ± 0%   658.2µ ± 0%  +41.87% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH4-8       374.8µ ± 1%   526.7µ ± 1%  +40.54% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH5-8       314.4µ ± 1%   439.7µ ± 0%  +39.86% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH6-8       288.1µ ± 2%   376.4µ ± 1%  +30.66% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH7-8       232.8µ ± 1%   331.2µ ± 0%  +42.30% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH8-8       265.7µ ± 0%   293.6µ ± 0%  +10.48% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH9-8       290.3µ ± 2%   264.9µ ± 1%   -8.74% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH10-8      394.0µ ± 0%   240.6µ ± 0%  -38.94% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH11-8      266.8µ ± 0%   221.2µ ± 0%  -17.08% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH12-8      274.6µ ± 3%   204.4µ ± 0%  -25.54% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH13-8      259.2µ ± 2%   189.8µ ± 0%  -26.75% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH14-8      296.1µ ± 2%   177.5µ ± 0%  -40.03% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH15-8      200.0µ ± 1%   166.4µ ± 0%  -16.79% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH16-8      155.3µ ± 0%   156.6µ ± 0%   +0.88% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH17-8      177.0µ ± 1%   148.0µ ± 0%  -16.35% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH18-8      227.9µ ± 1%   140.6µ ± 1%  -38.32% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH19-8      173.5µ ± 1%   133.7µ ± 1%  -22.96% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH20-8      166.1µ ± 3%   127.4µ ± 1%  -23.35% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH21-8      166.2µ ± 2%   121.8µ ± 1%  -26.71% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH22-8      188.0µ ± 2%   117.0µ ± 1%  -37.77% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH23-8      147.7µ ± 4%   111.8µ ± 1%  -24.30% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH24-8      139.7µ ± 2%   107.2µ ± 0%  -23.30% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH25-8      177.9µ ± 3%   103.5µ ± 1%  -41.84% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH26-8      171.1µ ± 4%   100.1µ ± 1%  -41.48% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH27-8     188.84µ ± 6%   96.44µ ± 2%  -48.93% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH28-8     163.66µ ± 4%   92.95µ ± 1%  -43.21% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH29-8     158.16µ ± 6%   89.74µ ± 0%  -43.26% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH30-8     184.84µ ± 3%   94.65µ ± 6%  -48.80% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH31-8     157.25µ ± 3%   84.39µ ± 2%  -46.33% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH32-8     122.27µ ± 2%   55.22µ ± 2%  -54.84% (p=0.000 n=10)
JumpdestOpAnalysis/JUMPDEST-8    562.5µ ± 0%   533.9µ ± 5%   -5.08% (p=0.007 n=10)
JumpdestOpAnalysis/STOP-8        569.8µ ± 2%   530.3µ ± 0%   -6.93% (p=0.000 n=10)
geomean                          217.3µ        171.9µ       -20.89%
```
<details>
<summary>Old Results</summary>
<pre>
$ benchstat -col /v bitvec.bench
                              │     old      │                 new                  │
                              │    sec/op    │    sec/op     vs base                │
JumpdestAnalysis_rand-8         1111.5µ ± 0%    508.5µ ± 1%  -54.25% (p=0.000 n=20)
JumpdestAnalysis_weth9-8        1006.0n ± 1%    808.0n ± 1%  -19.68% (p=0.000 n=20)
JumpdestOpAnalysis/PUSH1-8       844.5µ ± 3%   1550.8µ ± 1%  +83.65% (p=0.000 n=20)
JumpdestOpAnalysis/PUSH2-8       588.2µ ± 2%   1028.3µ ± 1%  +74.81% (p=0.000 n=20)
JumpdestOpAnalysis/PUSH3-8       459.3µ ± 1%    773.2µ ± 3%  +68.32% (p=0.000 n=20)
JumpdestOpAnalysis/PUSH4-8       425.8µ ± 2%    618.2µ ± 2%  +45.18% (p=0.000 n=20)
JumpdestOpAnalysis/PUSH5-8       349.6µ ± 1%    490.2µ ± 1%  +40.23% (p=0.000 n=20)
JumpdestOpAnalysis/PUSH6-8       322.1µ ± 1%    450.8µ ± 1%  +39.93% (p=0.000 n=20)
JumpdestOpAnalysis/PUSH7-8       233.2µ ± 3%    393.2µ ± 1%  +68.60% (p=0.000 n=20)
JumpdestOpAnalysis/PUSH8-8       268.2µ ± 5%    341.3µ ± 2%  +27.29% (p=0.000 n=20)
JumpdestOpAnalysis/PUSH9-8       309.1µ ± 1%    311.7µ ± 0%   +0.85% (p=0.009 n=20)
JumpdestOpAnalysis/PUSH10-8      410.0µ ± 4%    279.5µ ± 2%  -31.82% (p=0.000 n=20)
JumpdestOpAnalysis/PUSH11-8      278.7µ ± 3%    265.1µ ± 1%   -4.89% (p=0.000 n=20)
JumpdestOpAnalysis/PUSH12-8      309.4µ ± 1%    225.2µ ± 1%  -27.21% (p=0.000 n=20)
JumpdestOpAnalysis/PUSH13-8      287.4µ ± 1%    226.8µ ± 1%  -21.08% (p=0.000 n=20)
JumpdestOpAnalysis/PUSH14-8      292.1µ ± 2%    213.8µ ± 1%  -26.81% (p=0.000 n=20)
JumpdestOpAnalysis/PUSH15-8      203.7µ ± 5%    196.1µ ± 3%   -3.73% (p=0.000 n=20)
JumpdestOpAnalysis/PUSH16-8      168.8µ ± 1%    186.8µ ± 1%  +10.66% (p=0.000 n=20)
JumpdestOpAnalysis/PUSH17-8      190.1µ ± 1%    168.9µ ± 5%  -11.17% (p=0.000 n=20)
JumpdestOpAnalysis/PUSH18-8      251.5µ ± 1%    156.5µ ± 2%  -37.78% (p=0.000 n=20)
JumpdestOpAnalysis/PUSH19-8      185.0µ ± 3%    159.9µ ± 1%  -13.57% (p=0.000 n=20)
JumpdestOpAnalysis/PUSH20-8      170.4µ ± 1%    152.4µ ± 1%  -10.60% (p=0.000 n=20)
JumpdestOpAnalysis/PUSH21-8      173.1µ ± 6%    146.2µ ± 1%  -15.57% (p=0.000 n=20)
JumpdestOpAnalysis/PUSH22-8      212.5µ ± 1%    137.0µ ± 4%  -35.52% (p=0.000 n=20)
JumpdestOpAnalysis/PUSH23-8      156.2µ ± 1%    123.1µ ± 1%  -21.17% (p=0.000 n=20)
JumpdestOpAnalysis/PUSH24-8      151.3µ ± 1%    129.5µ ± 2%  -14.38% (p=0.000 n=20)
JumpdestOpAnalysis/PUSH25-8      181.8µ ± 3%    125.5µ ± 1%  -30.99% (p=0.000 n=20)
JumpdestOpAnalysis/PUSH26-8      176.9µ ± 3%    120.9µ ± 1%  -31.69% (p=0.000 n=20)
JumpdestOpAnalysis/PUSH27-8      199.0µ ± 1%    116.0µ ± 1%  -41.72% (p=0.000 n=20)
JumpdestOpAnalysis/PUSH28-8      188.5µ ± 1%    108.6µ ± 2%  -42.43% (p=0.000 n=20)
JumpdestOpAnalysis/PUSH29-8      182.2µ ± 1%    102.8µ ± 1%  -43.56% (p=0.000 n=20)
JumpdestOpAnalysis/PUSH30-8      184.1µ ± 1%    105.6µ ± 1%  -42.64% (p=0.000 n=20)
JumpdestOpAnalysis/PUSH31-8      172.3µ ± 3%    102.9µ ± 1%  -40.31% (p=0.000 n=20)
JumpdestOpAnalysis/PUSH32-8     124.65µ ± 1%    99.15µ ± 1%  -20.46% (p=0.000 n=20)
JumpdestOpAnalysis/JUMPDEST-8    552.5µ ± 0%    401.3µ ± 3%  -27.37% (p=0.000 n=20)
JumpdestOpAnalysis/STOP-8        580.9µ ± 1%    409.6µ ± 1%  -29.49% (p=0.000 n=20)
geomean                          228.9µ         199.4µ       -12.89%
</pre>
</details>