@holiman 2024-02-14
Hm, so the mean is faster, 

```
geomean                          228.9µ         199.4µ       -12.89%
``` 
But I don't think that's as important as keeping the worst-case in check. 
```
JumpdestOpAnalysis/PUSH1-8       844.5µ ± 3%   1550.8µ ± 1%  +83.65% (p=0.000 n=20)
```
To be safe against attacks, we need to ensure that the worst-case doesn't hurt us too much.