@lmittmann 2024-02-14
> But I don't think that's as important as keeping the worst-case in check.

I see your point. Although if you look at this the other way around you have the same worst-case somewhere else:
```
                            │     new     │                 old                 │
                            │   sec/op    │   sec/op     vs base                │
JumpdestOpAnalysis/PUSH27-8   116.0µ ± 1%   199.0µ ± 1%  +71.58% (p=0.000 n=20)
JumpdestOpAnalysis/PUSH28-8   108.6µ ± 2%   188.5µ ± 1%  +73.69% (p=0.000 n=20)
JumpdestOpAnalysis/PUSH29-8   102.8µ ± 1%   182.2µ ± 1%  +77.17% (p=0.000 n=20)
JumpdestOpAnalysis/PUSH30-8   105.6µ ± 1%   184.1µ ± 1%  +74.34% (p=0.000 n=20)
```

FWIW I see an overall VM performance improvement of 2% in my real-mainnet-blockinterval benchmark. This includes additional overhead of my `w3vm` package, so the pure `core/vm` performance improvement should be even slightly better (https://github.com/lmittmann/w3/blob/main/w3vm/bench_test.go).