@holiman 2024-02-14
A sequence of `PUSH1` is worst case:  `844.5µ < 116.0µ`. I don't understand how you can interpret it any other way ?

- A non-PUSH will shortcut the loop early (but also always do `n` spins in the loop for `n` ops)
- A HIGH push, like `PUSH32` will advance in steps of `32`, thus doing _way_ less spins in the loop
- A low push, like `PUSH1` will enter the heavy part of the loop every time, and although it will skip ahead, it will only skip ahead one step, so will do `n/2` spins in the loop for `n` ops).