@lmittmann 2024-02-14
Yes, I see your point, in absolute numbers `PUSH1` is clearly the worst case. Maybe I manage to fix the worst case. The overall gain of 2% is quite nice IMO.