@holiman 2024-02-14
Sure! I suspect that the map lookups are taking a toll.