@lmittmann 2024-02-14
> Sure! I suspect that the map lookups are taking a toll.
No, that's an array lookup and actually an optimization.