@lmittmann 2024-02-22
@holiman I updated the implementation and the benchmark results. The mean performance improved by another 8%. `PUSH1` is now slightly faster than the original implementation. The new worst-case is `PUSH2`. However, its performance is close to the performance of `PUSH1` in the old implementation.