@holiman 2024-03-12
Not yet, no. I'm curious about 
```
JumpdestAnalysis_rand-8         1133.2µ ± 1%   542.6µ ± 1%  -52.11% (p=0.000 n=10)
```
I don't see any reason for a 52% speedup in the general/random case. Sometimes there are just weird side-effects depending on certain options that the compiler happens to do, or not do, which don't really correspond to actual code features.