@lmittmann 2024-03-20
I don't really understand the degree of performance increase for the random case either. However, the overall performance improvement comes from the reduced number of slice accesses and thus the fewer number of slice bound checks inserted by the compiler. See https://go.godbolt.org/z/hrE5e7naP