@lmittmann 2024-04-14
hi @holiman, any chance of getting this performance improvement merged?