@lmittmann 2024-04-14
if the code is okay so far from your end I will finalize the PR and drop the duplicated code I kept for testing and benchmarking