@namiloh 2024-04-14
I'd drop it, yes, but: put all benchmarking changes in one commit, and code-changes in a second. That makes it easy to compare by just jumping forward/back