@lmittmann 2024-04-16
@holiman I dropped the old implementation for the new. The old benchmark logic was not changed to keep them comparable (I added a few new though). The test cases where translated from the old impl to the new one.

* old code with additional benchmarks: [04593bb](https://github.com/ethereum/go-ethereum/pull/28984/commits/04593bb71a6219a8560dd8e593709962339900b8)
* new code: [071bb70](https://github.com/ethereum/go-ethereum/pull/28984/commits/071bb7082732ab59e4684e3617953d4f6a470748)