@holiman 2024-04-17
Here are my numbers. 
The `1200k` is empty code, just `STOP` opcodes. It becomes 30% slower, whereas the totally random data becomes almost twice as fast. The weth-code drops 30% performance. 

```
name                      old time/op    new time/op    delta
JumpdestAnalysis_1200k-8    1.19ms ±14%    1.71ms ± 7%  +44.30%  (p=0.000 n=10+9)
JumpdestAnalysis_rand-8     2.90ms ±13%    1.53ms ±15%  -47.45%  (p=0.000 n=10+10)
JumpdestAnalysis_weth9-8    2.34µs ±57%    3.10µs ±31%  +32.24%  (p=0.029 n=10+10)

name                      old speed      new speed      delta
JumpdestAnalysis_1200k-8  1.04GB/s ±16%  0.72GB/s ± 8%  -31.05%  (p=0.000 n=10+9)
JumpdestAnalysis_rand-8    425MB/s ±12%   814MB/s ±14%  +91.50%  (p=0.000 n=10+10)
JumpdestAnalysis_weth9-8  1.43GB/s ±41%  1.04GB/s ±41%  -27.28%  (p=0.029 n=10+10)
```
As for the detailed opcodes: 
```
user@work go-ethereum]$ benchstat analysis.ops.1  analysis.ops.2
name                           old time/op    new time/op    delta
JumpdestOpAnalysis/PUSH1-8       2.43ms ±44%    2.16ms ±30%      ~     (p=0.529 n=10+10)
JumpdestOpAnalysis/PUSH2-8       1.95ms ±26%    2.67ms ±29%   +36.85%  (p=0.009 n=10+10)
JumpdestOpAnalysis/PUSH3-8       1.66ms ±44%    1.91ms ±27%      ~     (p=0.436 n=10+10)
JumpdestOpAnalysis/PUSH4-8       1.25ms ±28%    1.58ms ±24%      ~     (p=0.063 n=10+10)
JumpdestOpAnalysis/PUSH5-8        974µs ±31%    1363µs ±32%   +39.89%  (p=0.007 n=10+10)
JumpdestOpAnalysis/PUSH6-8        905µs ±22%    1090µs ±26%      ~     (p=0.052 n=10+10)
JumpdestOpAnalysis/PUSH7-8        828µs ±24%     915µs ±32%      ~     (p=0.631 n=10+10)
JumpdestOpAnalysis/PUSH8-8        948µs ±27%     721µs ±12%   -23.98%  (p=0.010 n=10+9)
JumpdestOpAnalysis/PUSH9-8        935µs ±27%     759µs ±27%   -18.84%  (p=0.035 n=10+10)
JumpdestOpAnalysis/PUSH10-8      1.14ms ± 0%    0.74ms ±30%   -34.91%  (p=0.000 n=8+10)
JumpdestOpAnalysis/PUSH11-8       885µs ±36%     610µs ±32%   -31.11%  (p=0.005 n=10+10)
JumpdestOpAnalysis/PUSH12-8       846µs ±27%     547µs ±33%   -35.28%  (p=0.000 n=10+10)
JumpdestOpAnalysis/PUSH13-8       794µs ±33%     586µs ±29%   -26.12%  (p=0.011 n=10+10)
JumpdestOpAnalysis/PUSH14-8       883µs ± 1%     479µs ±36%   -45.75%  (p=0.000 n=9+10)
JumpdestOpAnalysis/PUSH15-8       811µs ± 2%     527µs ±26%   -35.02%  (p=0.000 n=8+10)
JumpdestOpAnalysis/PUSH16-8       520µs ±29%     484µs ±26%      ~     (p=0.089 n=10+10)
JumpdestOpAnalysis/PUSH17-8       520µs ±33%     455µs ±29%      ~     (p=0.218 n=10+10)
JumpdestOpAnalysis/PUSH18-8       580µs ±24%     420µs ±31%   -27.67%  (p=0.035 n=10+10)
JumpdestOpAnalysis/PUSH19-8       461µs ± 7%     489µs ± 0%    +6.01%  (p=0.010 n=8+8)
JumpdestOpAnalysis/PUSH20-8       567µs ±28%     411µs ±27%   -27.53%  (p=0.015 n=10+10)
JumpdestOpAnalysis/PUSH21-8       562µs ±26%     351µs ±27%   -37.53%  (p=0.000 n=10+10)
JumpdestOpAnalysis/PUSH22-8       610µs ± 1%     383µs ±31%   -37.23%  (p=0.000 n=8+10)
JumpdestOpAnalysis/PUSH23-8       480µs ±22%     333µs ±27%   -30.57%  (p=0.004 n=10+10)
JumpdestOpAnalysis/PUSH24-8       538µs ±31%     342µs ±27%   -36.43%  (p=0.000 n=10+10)
JumpdestOpAnalysis/PUSH25-8       545µs ±32%     280µs ±36%   -48.62%  (p=0.000 n=10+10)
JumpdestOpAnalysis/PUSH26-8       615µs ±23%     283µs ±30%   -54.04%  (p=0.000 n=10+10)
JumpdestOpAnalysis/PUSH27-8       426µs ±12%     278µs ±28%   -34.87%  (p=0.000 n=10+10)
JumpdestOpAnalysis/PUSH28-8       506µs ±27%     268µs ±28%   -46.98%  (p=0.000 n=10+10)
JumpdestOpAnalysis/PUSH29-8       503µs ±32%     277µs ±30%   -44.96%  (p=0.000 n=10+10)
JumpdestOpAnalysis/PUSH30-8       429µs ±40%     271µs ±30%   -36.93%  (p=0.000 n=10+10)
JumpdestOpAnalysis/PUSH31-8       447µs ±30%     310µs ± 1%   -30.57%  (p=0.000 n=10+8)
JumpdestOpAnalysis/PUSH32-8       407µs ±39%     191µs ±37%   -53.20%  (p=0.000 n=10+10)
JumpdestOpAnalysis/JUMPDEST-8    1.34ms ±53%    1.55ms ±42%      ~     (p=0.105 n=10+10)
JumpdestOpAnalysis/STOP-8        1.08ms ±52%    1.43ms ±40%   +31.85%  (p=0.029 n=10+10)

name                           old speed      new speed      delta
JumpdestOpAnalysis/PUSH1-8      545MB/s ±65%   602MB/s ±28%      ~     (p=0.529 n=10+10)
JumpdestOpAnalysis/PUSH2-8      659MB/s ±30%   481MB/s ±34%   -26.97%  (p=0.009 n=10+10)
JumpdestOpAnalysis/PUSH3-8      839MB/s ±39%   664MB/s ±32%      ~     (p=0.436 n=10+10)
JumpdestOpAnalysis/PUSH4-8     1.04GB/s ±32%  0.80GB/s ±27%      ~     (p=0.063 n=10+10)
JumpdestOpAnalysis/PUSH5-8     1.32GB/s ±28%  0.95GB/s ±40%   -28.09%  (p=0.007 n=10+10)
JumpdestOpAnalysis/PUSH6-8     1.39GB/s ±20%  1.17GB/s ±25%      ~     (p=0.052 n=10+10)
JumpdestOpAnalysis/PUSH7-8     1.53GB/s ±29%  1.40GB/s ±28%      ~     (p=0.631 n=10+10)
JumpdestOpAnalysis/PUSH8-8     1.35GB/s ±32%  1.66GB/s ±31%   +23.05%  (p=0.019 n=10+10)
JumpdestOpAnalysis/PUSH9-8     1.37GB/s ±32%  1.70GB/s ±26%   +24.40%  (p=0.035 n=10+10)
JumpdestOpAnalysis/PUSH10-8    1.08GB/s ± 0%  1.75GB/s ±35%   +62.48%  (p=0.000 n=8+10)
JumpdestOpAnalysis/PUSH11-8    1.46GB/s ±30%  2.08GB/s ±27%   +42.38%  (p=0.005 n=10+10)
JumpdestOpAnalysis/PUSH12-8    1.50GB/s ±32%  2.29GB/s ±26%   +52.68%  (p=0.000 n=10+10)
JumpdestOpAnalysis/PUSH13-8    1.62GB/s ±42%  2.18GB/s ±35%   +34.11%  (p=0.011 n=10+10)
JumpdestOpAnalysis/PUSH14-8    1.39GB/s ± 1%  2.70GB/s ±30%   +93.92%  (p=0.000 n=9+10)
JumpdestOpAnalysis/PUSH15-8    1.51GB/s ± 2%  2.41GB/s ±31%   +59.46%  (p=0.000 n=8+10)
JumpdestOpAnalysis/PUSH16-8    2.49GB/s ±35%  2.65GB/s ±29%      ~     (p=0.089 n=10+10)
JumpdestOpAnalysis/PUSH17-8    2.54GB/s ±40%  2.82GB/s ±36%      ~     (p=0.218 n=10+10)
JumpdestOpAnalysis/PUSH18-8    2.20GB/s ±26%  3.07GB/s ±38%   +39.20%  (p=0.035 n=10+10)
JumpdestOpAnalysis/PUSH19-8    2.67GB/s ± 7%  2.51GB/s ± 0%    -5.88%  (p=0.010 n=8+8)
JumpdestOpAnalysis/PUSH20-8    2.25GB/s ±34%  3.09GB/s ±32%   +37.20%  (p=0.015 n=10+10)
JumpdestOpAnalysis/PUSH21-8    2.26GB/s ±32%  3.65GB/s ±31%   +61.63%  (p=0.000 n=10+10)
JumpdestOpAnalysis/PUSH22-8    2.01GB/s ± 1%  3.33GB/s ±39%   +65.54%  (p=0.000 n=8+10)
JumpdestOpAnalysis/PUSH23-8    2.64GB/s ±22%  3.88GB/s ±30%   +46.75%  (p=0.004 n=10+10)
JumpdestOpAnalysis/PUSH24-8    2.35GB/s ±41%  3.72GB/s ±32%   +58.29%  (p=0.000 n=10+10)
JumpdestOpAnalysis/PUSH25-8    2.37GB/s ±40%  4.55GB/s ±29%   +92.15%  (p=0.000 n=10+10)
JumpdestOpAnalysis/PUSH26-8    2.04GB/s ±27%  4.55GB/s ±29%  +123.64%  (p=0.000 n=10+10)
JumpdestOpAnalysis/PUSH27-8    2.90GB/s ±12%  4.65GB/s ±28%   +60.77%  (p=0.000 n=10+10)
JumpdestOpAnalysis/PUSH28-8    2.51GB/s ±25%  4.81GB/s ±29%   +92.06%  (p=0.000 n=10+10)
JumpdestOpAnalysis/PUSH29-8    2.59GB/s ±38%  4.65GB/s ±36%   +79.68%  (p=0.000 n=10+10)
JumpdestOpAnalysis/PUSH30-8    3.04GB/s ±33%  4.73GB/s ±37%   +55.78%  (p=0.000 n=10+10)
JumpdestOpAnalysis/PUSH31-8    2.92GB/s ±35%  3.96GB/s ± 1%   +35.77%  (p=0.000 n=10+8)
JumpdestOpAnalysis/PUSH32-8    3.22GB/s ±54%  6.85GB/s ±49%  +112.96%  (p=0.000 n=10+10)
JumpdestOpAnalysis/JUMPDEST-8  1.06GB/s ±84%  0.89GB/s ±52%      ~     (p=0.105 n=10+10)
JumpdestOpAnalysis/STOP-8      1.25GB/s ±41%  0.94GB/s ±42%   -24.93%  (p=0.029 n=10+10)
```

The new worst-case is `PUSH2`, which is worse than the previous worstcase `PUSH1`. 

This PR has the drawback that it makes the memory consumption larger (esp for small codes), since it allocates `32 * len(code)/32+2)` instead of `len(code)/8+1+4)` bytes. So `64` bytes for a 1-byte contract, instead of `5`. 

This PR has the upside that it's a bit simpler code. But OTOH it's new code, whereas the existing is old and trusted. 

All in all, I'm not sure if it's worth it. Feels a bit like whack-a-mole with speeds, improve it a bit here and it becomes slower there. I am not convinced that this PR is overall an improvement.