@lmittmann 2024-04-24
Interesting, you have completely different results than I do. On which CPU architecture are you?

I just reran the benchmarks at the two commits, and essentially PUSH1..8, and PUSH16 perform worse, everything else is better.

```
goos: linux
goarch: amd64
pkg: github.com/ethereum/go-ethereum/core/vm
cpu: 11th Gen Intel(R) Core(TM) i7-1185G7 @ 3.00GHz
                              │   old.bench   │              new.bench              │
                              │    sec/op     │   sec/op     vs base                │
JumpdestAnalysis_1200k-8         568.1µ ±  0%   537.5µ ± 0%   -5.38% (p=0.000 n=10)
JumpdestAnalysis_rand-8         1029.6µ ±  0%   519.6µ ± 3%  -49.53% (p=0.000 n=10)
JumpdestAnalysis_weth9-8        1036.5n ±  1%   959.9n ± 4%   -7.39% (p=0.000 n=10)
JumpdestHashing_1200k-8          2.740m ±  3%   2.709m ± 5%        ~ (p=0.247 n=10)
JumpdestOpAnalysis/PUSH1-8       796.0µ ±  1%   838.9µ ± 3%   +5.38% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH2-8       607.8µ ±  3%   980.6µ ± 1%  +61.35% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH3-8       453.3µ ±  3%   734.6µ ± 3%  +62.05% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH4-8       443.8µ ± 10%   622.6µ ± 4%  +40.27% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH5-8       354.2µ ±  2%   529.5µ ± 2%  +49.51% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH6-8       302.0µ ±  1%   447.3µ ± 2%  +48.12% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH7-8       249.0µ ±  1%   397.1µ ± 5%  +59.50% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH8-8       256.6µ ±  1%   328.7µ ± 2%  +28.09% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH9-8       309.1µ ±  1%   301.9µ ± 2%   -2.32% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH10-8      404.1µ ±  1%   272.2µ ± 1%  -32.64% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH11-8      284.8µ ±  1%   248.3µ ± 1%  -12.84% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH12-8      289.3µ ±  1%   233.9µ ± 4%  -19.17% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH13-8      261.7µ ±  1%   229.7µ ± 2%  -12.23% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH14-8      247.2µ ±  2%   217.2µ ± 2%  -12.14% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH15-8      213.9µ ±  1%   200.2µ ± 2%   -6.41% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH16-8      151.8µ ±  1%   191.5µ ± 1%  +26.14% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH17-8      184.9µ ±  2%   181.5µ ± 2%   -1.88% (p=0.019 n=10)
JumpdestOpAnalysis/PUSH18-8      246.0µ ±  1%   165.0µ ± 2%  -32.94% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH19-8      177.5µ ±  3%   151.6µ ± 3%  -14.56% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH20-8      188.6µ ±  4%   142.6µ ± 2%  -24.42% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH21-8      184.0µ ±  2%   136.8µ ± 1%  -25.63% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH22-8      211.5µ ±  1%   134.4µ ± 7%  -36.48% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH23-8      156.7µ ±  1%   139.4µ ± 1%  -11.07% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH24-8      152.4µ ±  1%   133.9µ ± 2%  -12.11% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH25-8      190.5µ ±  1%   128.1µ ± 1%  -32.75% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH26-8      185.3µ ±  1%   123.0µ ± 1%  -33.60% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH27-8      184.0µ ±  1%   120.0µ ± 2%  -34.79% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH28-8      181.8µ ±  2%   114.0µ ± 1%  -37.31% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH29-8      172.5µ ±  1%   105.4µ ± 3%  -38.89% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH30-8      168.6µ ±  2%   100.2µ ± 2%  -40.59% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH31-8      158.9µ ± 10%   102.3µ ± 3%  -35.61% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH32-8     129.09µ ±  1%   64.81µ ± 2%  -49.79% (p=0.000 n=10)
JumpdestOpAnalysis/JUMPDEST-8    580.7µ ±  2%   545.4µ ± 2%   -6.08% (p=0.000 n=10)
JumpdestOpAnalysis/STOP-8        581.0µ ±  1%   551.8µ ± 2%   -5.03% (p=0.000 n=10)
geomean                          246.7µ         218.4µ       -11.49%

                              │   old.bench   │               new.bench               │
                              │      B/s      │      B/s       vs base                │
JumpdestAnalysis_1200k-8        2.014Gi ±  0%    2.129Gi ± 0%   +5.69% (p=0.000 n=10)
JumpdestAnalysis_rand-8         1.112Gi ±  0%    2.202Gi ± 3%  +98.13% (p=0.000 n=10)
JumpdestAnalysis_weth9-8        2.806Gi ±  1%    3.031Gi ± 4%   +8.02% (p=0.000 n=10)
JumpdestHashing_1200k-8         427.9Mi ±  3%    432.6Mi ± 5%        ~ (p=0.247 n=10)
JumpdestOpAnalysis/PUSH1-8      1.438Gi ±  1%    1.364Gi ± 3%   -5.10% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH2-8      1.883Gi ±  3%    1.167Gi ± 1%  -38.02% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH3-8      2.525Gi ±  3%    1.558Gi ± 2%  -38.29% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH4-8      2.578Gi ± 11%    1.839Gi ± 4%  -28.68% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH5-8      3.231Gi ±  2%    2.161Gi ± 2%  -33.11% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH6-8      3.789Gi ±  1%    2.558Gi ± 2%  -32.49% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH7-8      4.596Gi ±  1%    2.882Gi ± 5%  -37.31% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH8-8      4.460Gi ±  1%    3.482Gi ± 2%  -21.93% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH9-8      3.703Gi ±  1%    3.791Gi ± 2%   +2.37% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH10-8     2.832Gi ±  1%    4.204Gi ± 1%  +48.45% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH11-8     4.018Gi ±  1%    4.610Gi ± 1%  +14.73% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH12-8     3.956Gi ±  1%    4.894Gi ± 4%  +23.71% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH13-8     4.372Gi ±  1%    4.981Gi ± 2%  +13.93% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH14-8     4.629Gi ±  2%    5.268Gi ± 2%  +13.82% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH15-8     5.349Gi ±  1%    5.716Gi ± 2%   +6.86% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH16-8     7.538Gi ±  1%    5.976Gi ± 1%  -20.72% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH17-8     6.188Gi ±  2%    6.306Gi ± 2%   +1.91% (p=0.019 n=10)
JumpdestOpAnalysis/PUSH18-8     4.651Gi ±  1%    6.936Gi ± 2%  +49.11% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH19-8     6.448Gi ±  3%    7.547Gi ± 3%  +17.03% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH20-8     6.067Gi ±  4%    8.026Gi ± 2%  +32.30% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH21-8     6.219Gi ±  2%    8.363Gi ± 1%  +34.47% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH22-8     5.410Gi ±  1%    8.518Gi ± 7%  +57.44% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH23-8     7.302Gi ±  1%    8.211Gi ± 1%  +12.45% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH24-8     7.509Gi ±  1%    8.544Gi ± 2%  +13.78% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH25-8     6.007Gi ±  1%    8.933Gi ± 1%  +48.71% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH26-8     6.177Gi ±  1%    9.302Gi ± 1%  +50.61% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH27-8     6.218Gi ±  1%    9.535Gi ± 2%  +53.35% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH28-8     6.295Gi ±  2%   10.042Gi ± 1%  +59.51% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH29-8     6.635Gi ±  2%   10.859Gi ± 3%  +63.65% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH30-8     6.788Gi ±  2%   11.426Gi ± 2%  +68.32% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH31-8     7.204Gi ±  9%   11.188Gi ± 3%  +55.30% (p=0.000 n=10)
JumpdestOpAnalysis/PUSH32-8     8.866Gi ±  1%   17.658Gi ± 2%  +99.18% (p=0.000 n=10)
JumpdestOpAnalysis/JUMPDEST-8   1.971Gi ±  2%    2.098Gi ± 2%   +6.47% (p=0.000 n=10)
JumpdestOpAnalysis/STOP-8       1.970Gi ±  1%    2.074Gi ± 2%   +5.30% (p=0.000 n=10)
geomean                         3.963Gi          4.478Gi       +12.98%
```